requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==5.4.0
MarkupSafe==3.0.2
more-itertools==10.7.0
multidict==6.4.4
narwhals==1.41.0
numpy==2.2.6
openai==0.28.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.1
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rpds-py==0.25.1
simple-salesforce==1.12.6
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
yarl==1.20.0
zeep==4.3.1
